Divide usage.py into a script and function modules. Modify README to reflect new changes to code structure.
